Fix Login: Application crashes on logging in from logout page
on Logout application navigates to '/logout' and it makes api request,then it will show login form.But in
login component, after a successfull login it will redirect to previous screen,In this case '/logout' screen.
because we used componentDidRecieveProp, it will be invoked. this process forms a continuose loop.as the result
application gets crash. This has been resolved by navigating to a different screen '/' after login.

Fixes: #1147