build(devel): Avoid real file interfere w/ target
By default `./Makefile` targets are expected to create or update an actual file or
directory build artifact and use its modification time (compared to the modification
times of its prerequisites) to determine if the recipe for that target needs to be run.
As such, if a target isn't intended to actually produce a build artifact, but instead to
be used merely as a shorthand for running the recipe, it can be interfered with if that
file ever comes to exist.  Usually that means it just stops doing anything and `$ make`
won't give you any clues as to why.  Best to declare them specifically with `.PHONY:
...`.

It's also better to put the `.PHONY: ...` declarations with each target.  I've never
seen a case where a single `.PHONY: ...` declaration at the end of the `./Makefile` was
used where it wasn't perpetually getting out of data/sync with the actual targets.