build(devel): Improve browsing API source code
Use `collective.recipe.omelette` to reproduce a single directory tree of the Python
packages installed in the `./api/` buildout's `instance` part.  Useful for searching,
browsing, or otherwise exploring all the source code involved in the application in a
way that's more readable and avoids duplicates from older versions of eggs.