refactor(user): Factor user session state to selectors
Refactor the logic around user session state out of the multiple components into Redux
selectors.  This is a step on the way to refactoring the UI user session logic to be
agnostic of *how* the user is authenticated.  Currently this includes 2 selectors:

- detecting whether a user is currently authenticated and logged in
- retrieving the authenticated users login or ID

I note this project isn't using selectors prior to this.  From [the Redux
docs](https://redux.js.org/tutorials/fundamentals/part-2-concepts-data-flow#selectors):

> Selectors are functions that know how to extract specific pieces of information from a
> store state value. As an application grows bigger, this can help avoid repeating logic
> as different parts of the app need to read the same data

That's *exactly* what's needed here.  This is particularly true for getting user data
state, such as the `userId`.  As such, I went ahead and introduced the usage of
selectors here.

Refs #134784