refactor(actions): Easier action lookup by id
Seeking an action by id, for example to check for the presence of an action or to lookup
the `url` property of a specific action, is a very common task.  As such, transform the
arrays of actions into objects of actions by action id and include that in Redux state
as well.

Next, I plan to use this to replace the JWT implementation-dependent check if a user is
logged in.

Refs #134784