fix(auth): UI logged-in detection for other auths (#2535)
* build(devel): Add a proxy hosting testbed

A testbed that reproduces accessing all of the ZMI, the classic HTML plone UI, and the
Volto UI under the same hostname.

Originally needed to workout issues with different kinds of authentication working for
all UIs.

refs #131171

* build(devel): Improve browsing API source code

Use `collective.recipe.omelette` to reproduce a single directory tree of the Python
packages installed in the `./api/` buildout's `instance` part.  Useful for searching,
browsing, or otherwise exploring all the source code involved in the application in a
way that's more readable and avoids duplicates from older versions of eggs.

* build(devel): Wrong checkout default branch name

This repository has no `master` branch.  At least according to GitHub, `main` is the
default branch and it also has the most recent activity in `$ git log`.

* docs(redux): Order folder structure per frameworks

More sensible ordering of the folder structure documentation so that the library, React,
is first, the state framework, Redux, is next, followed by more project specific
structure.

* refactor(user): Factor user session state to selectors

Refactor the logic around user session state out of the multiple components into Redux
selectors.  This is a step on the way to refactoring the UI user session logic to be
agnostic of *how* the user is authenticated.  Currently this includes 2 selectors:

- detecting whether a user is currently authenticated and logged in
- retrieving the authenticated users login or ID

I note this project isn't using selectors prior to this.  From [the Redux
docs](https://redux.js.org/tutorials/fundamentals/part-2-concepts-data-flow#selectors):

> Selectors are functions that know how to extract specific pieces of information from a
> store state value. As an application grows bigger, this can help avoid repeating logic
> as different parts of the app need to read the same data

That's *exactly* what's needed here.  This is particularly true for getting user data
state, such as the `userId`.  As such, I went ahead and introduced the usage of
selectors here.

Refs #134784

* refactor(actions): Easier action lookup by id

Seeking an action by id, for example to check for the presence of an action or to lookup
the `url` property of a specific action, is a very common task.  As such, transform the
arrays of actions into objects of actions by action id and include that in Redux state
as well.

Next, I plan to use this to replace the JWT implementation-dependent check if a user is
logged in.

Refs #134784

* fix(auth): UI logged-in detection for other auths

Defer to the API, specifically the presence of the `login` action, to determine if a
user is logged in instead of depending on the implementation detail of the UI JWT login
process.

Refs #134784

* Fix prettier issues

* build(semver): React props changes break custom

Switch from a feature to a breaking change in the changelog
[because](https://github.com/plone/volto/pull/2535#issuecomment-875748237):

> In the RenderedView, the `token` is removed and used `userLoggedIn` instead. This will
> potentially break all customizations using component shadowing for all customized
> Views outthere, so we need to mark it as breaking.

* build(devel): Isolate Trafik proxy make testbed

Addresses [feedback](https://github.com/plone/volto/pull/2535#issuecomment-875748237):

> I would like to maintain the purity of what means to start the backend, what means to
> start the frontend without abstractions (docker in the middle using traeffic). It's
> fine to maintain it as a separate command, but not as "canonical". The reasons are
> mainly for training and newbies to really know what's going on under the hood.

Co-authored-by: Víctor Fernández de Alba <sneridagh@gmail.com>
Co-authored-by: Timo Stollenwerk <tisto@users.noreply.github.com>