build(proxy): Improve proxy testbed usage
When revisiting the proxy test bed I couldn't remember how to start it correctly.
Clearly I should have documented it better!  :-/ Clarify and expand the comments to
address that.

Also improves how the proxy test bed is run as well as the output from that test bed.

I change the target names from `start-*` to `run-*` because to my mind `start` means to
launch something successfully "in the background" and return the shell to the user, such
as with `# systemctl ...` and other system service control commands.  Whereas `run`
means to keep running "in the foreground" until the user interrupts.  This means the
terminology in the target names is inconsistent but I figured it was better to use
incrementally better terminology in the targets I'm touching and avoid breaking things
for other developers than to not improve the terminology at all.