docs(contrib): Changelog entry CI check exception (#2904)
The GitHub CI check for changelog entries misleads new, would be contributors
[indicating that entries should be added in `./news/*`
files](https://jenkins.plone.org/roboto/missing-changelog).

[Apparently](https://github.com/plone/volto/pull/2901#issuecomment-989581010) that check
comes from [mr.roboto](https://docs.plone.org/develop/coredev/docs/roboto.html).  If
that check successfully handles `./CHANGELOG.md` handles in addition to `towncrier`
`./news/*` entries, then it's failure message should reflect which is the case for the
repo it's checking.  As such, this does seem like a bug in `mr.roboto` or whatever code
and/or configuration is responsible for that check and its output.

[This project instead uses simple entries added to
`./CHANGELOG.md`](https://github.com/plone/volto/pull/2901#issuecomment-990856372) but
perhaps it should change to `./news/*` entries if only to be consistent with the rest of
the ecosystem to make contributing across the ecosystem a smoother experience.  In the
meantime, we should at least document this exception.