Fix select widget infinite loop
If an attempt is made to update a selection item that is not in the
choices, an infinite loop occurs. This can be caused by rare conditions
of a stale browser state.

This infinite loop has to be guarded by a condition in setState to
avoid triggering another component render and thus breaking the infinite
loop.