build(version): Improve version conflict resolution
I'm now getting a different version constraint conflict than I got before:

    Version and requirements information containing prompt-toolkit:
      [versions] constraint on prompt-toolkit: 1.0.18
      Requirement of robotframework-debuglibrary: prompt-toolkit<3,>=2
    While:
      Updating test.
    Error: The requirement ('prompt-toolkit<3,>=2') is not allowed by your [versions] constraint (1.0.18)

This is probably the better resolution anyways as it restores the primacy of the
"real" constraint from actual package requirements metadata over the version pins from
the base Plone buildout profile.  I think setting requirements to blank with `foo = `
used to accomplish this very thing, so I'm not sure why it's not working any more (or
right now), but addressing that is outside the scope of this PR and this gets us going
in the maentime.