feat(auth): Unified log in/out both API/classic
Revert the commits that changed the components that refer to the logged in user to use
the response from the actions API endpoint.  IOW, those components now, once again, use
the Bearer token instead.  Preserves all the other improvements made to the code along
the way.

The new branch for the `plone.restapi` includes the API changes to the PAS plugin needed
to set both the Plone classic session cookie and the Bearer token cookie that Volto can
optionally use to initialize authentication without using the Volto login component.

All this put together means that:
- the Volto components related to logged in users now behave largely as they did before
  in terms of interpreting state
- logging in to or out of either classic or Volto also does so in the other

One edge case that remains is when logging in via the Volto UI component as a user from
the Zope root `acl_users` (e.g. `admin` or `SITE_OWNER_NAME`), they aren't logged into
Plone classic.  The other direction is fine, logging in as `admin` to Plone classic
results in a new request to the Volto UI being logged in.  Digging into that now.

These changes do not address any of the issues that happen when logged in via the ZMI at
the Zope root.  A user logged in that way will still appear logged in to classic Plone,
but won't be recognized by Volto components as logged in but API responses will still
reflect their `Manager` access rights.  Tackling that is coming up.

This reverts commit ca0b83ffe50b18a16ffb67c0f1c5597ec2dfc16a.
This reverts commit faec2aedc6a815849a7642c6fb3b09b521f8a594.