build(checkouts): Easier development Plone sources
I kept going to Plone`s `buildout.coredev/sources.cfg` to add new checkouts, such as
while developing upstream PRs in combination with work here or after upstream PRs have
merged but not yet released.  It was extra work because I had to copy and paste the
`sources` line, but then manually replace the `${remotes:...}` substitutions.

This buildout is for development, we already use external `buildout:extends` URLs, so
would we repeat ourselves.