build(deps): Constrain, pin, upgrade API versions (#2922)
* build(make): Reduce noise for doc/help target

This is a regression from when I added defensive make settings in:

    build(devel): Avoid silent make errors/failures

* test(api): Restore test runner, ref'ed in Makefile

The `./Makefile` originally referenced a `test` target, since moved to the
`./api/Makefile`, but the `./api/buildout.cfg` installs no such buildout part.  Figuring
it's better to restore that than remove it, I tried restoring that part to the API
buildout and found it easy and that most of the tests are already passing.

* build(make): Reduce build times, incremental

For each step in the build process, use the real artifacts that result as targets in
`./api/Makefile`.  This prevents wasting time in the inner loop of individual
development and when the prerequisites, `... ./buildout.cfg ./src/*/setup.py`, are
maintained carefully this rarely results in surprising behavior such as something
failing because a build step wasn't detected as needed.

* build(checkout): Cleanup merged branch reference

* build(version): Fix version constraint conflicts

These popped up after merging master.  The conflicting version messages in order
encountered are:

    Version and requirements information containing plone.app.vocabularies:
      [versions] constraint on plone.app.vocabularies: 4.3.0 # plone.volto requires it
      Requirement of plone.volto: plone.app.vocabularies>=4.3.0
    While:
      Updating instance.
    Error: The requirement ('plone.app.vocabularies>=4.3.0') is not allowed by your
    [versions] constraint (4.3.0 # plone.volto requires it)

then:

    Version and requirements information containing plone.restapi:
      [versions] constraint on plone.restapi: 8.16.2 # force to latest p.restapi
      Requirement of plone.volto: plone.restapi>=8.13.0
      Requirement of Plone==5.2.6: plone.restapi
    While:
      Installing instance.
    Error: The requirement ('plone.restapi>=8.13.0') is not allowed by your [versions] constraint (8.16.2 # force to latest p.restapi)

then:

    Version and requirements information containing plone.rest:
      [versions] constraint on plone.rest: 2.0.0a1 # new JSON only traversal
      Requirement of plone.volto: plone.restapi>=8.13.0
      Requirement of Plone==5.2.6: plone.restapi
      Requirement of plone.restapi>=8.13.0: setuptools
      Requirement of plone.restapi>=8.13.0: pytz
      Requirement of plone.restapi>=8.13.0: python-dateutil
      Requirement of plone.restapi>=8.13.0: plone.schema>=1.2.1
      Requirement of plone.restapi>=8.13.0: plone.rest>=1.0a6
      Requirement of plone.restapi>=8.13.0: plone.behavior>=1.1
      Requirement of plone.restapi>=8.13.0: PyJWT
    While:
      Installing instance.
    Error: The requirement ('plone.rest>=1.0a6') is not allowed by your [versions] constraint (2.0.0a1 # new JSON only traversal)

* build(version): Improve version conflict resolution

I'm now getting a different version constraint conflict than I got before:

    Version and requirements information containing prompt-toolkit:
      [versions] constraint on prompt-toolkit: 1.0.18
      Requirement of robotframework-debuglibrary: prompt-toolkit<3,>=2
    While:
      Updating test.
    Error: The requirement ('prompt-toolkit<3,>=2') is not allowed by your [versions] constraint (1.0.18)

This is probably the better resolution anyways as it restores the primacy of the
"real" constraint from actual package requirements metadata over the version pins from
the base Plone buildout profile.  I think setting requirements to blank with `foo = `
used to accomplish this very thing, so I'm not sure why it's not working any more (or
right now), but addressing that is outside the scope of this PR and this gets us going
in the maentime.

* build(deps): Constrain, pin, upgrade API versions

Avoid misleading CI errors/failures, real user issues, etc. when new versions of
dependencies/requirements are released or upstream dependencies or constraints change.

Use [existing support in `zc.buildout` for managing
versions](http://www.buildout.org/en/latest/getting-started.html?highlight=update-versions-file#pinned-versions-1)
to capture version constraints required by this buildout that override the base Plone
version pins and automatically update the version pins in a separate file for
reproduciblity.  With these changes, as developers change requirements and/or
constraints, the pins in `./versions.cfg` should be updated when `$ ./bin/buildout` is
run and thus signal version changes to developers in VCS status.  These changes also add
a `$ make upgrade-versions` target to capture how to update all requirements to the
latest versions.

Note that since this changes version management and pins versions that weren't
previously pinned (a good thing) there's some risk this will result in CI
errors/failures that aren't necessarily related to these changes.  IOW, these changes
may reveal some hidden existing issues.  When these changes are merged, similar
unrelated CI errors/failures may be revealed in other branches.  IOW, there's some risk
this could be disruptive to the development community.  I can't think of any specific
way that would happen, CI should already be using the most recent versions, but it's
important to consider.

* build(clean): Fix clean checkout build issues

Also adds a target for running clean build tests via git worktrees going forward.

* build(make): Fix line continuation error

The extra newline without the line continuation had `$ make ...` treating the
prerequisites as a target recipe command.  Thanks to @sneridagh for the outsourced clean
build test that exposed this.  ;-)

* build(api): Minor performance, redundant install

Co-authored-by: Víctor Fernández de Alba <sneridagh@gmail.com>