Fix drag handler alignment for screen width < largeMonitorBreakpoint
Below 1224px, Both the drag handler and the delete button were pushed
inside the block editing border. This fix places them back outside
like they should be in all resolutions. Making sure that alignment is
the same both with normal blocks and the title block.

Note that the main part of this fix is fixing an override that was not
doing the override, but when I was there I also adjusted both icons in
all resolutions correctly.