fix iframe covering the page due to a react-error-overlay bug (#4242)
In devmode, nothing is clickable on the site. This occurs due to a bug
in react-error-overlay that displays an iframe that covers the whole
page and steals the focus. At the same time, there is an error appearing
in the console:

`ReferenceError: process is not defined`

The solution is to upgrade react-error-overlay.

This issue is [referenced here]
(https://stackoverflow.com/questions/70368760/react-uncaught-referenceerror-process-is-not-defined).
Also described in details
[here](https://github.com/facebook/create-react-app/issues/11773).