Remove the content encoding from the sitemap.xml.gz
Although it's all the same for Google with or without this, it's more
correct not to add the content encoding gzip header, as we just want
to transfer the gzipped file as a binary and not consider the browser
to decode it when downloaded.

(The other option would be to leave the content encoding header but then
just call the file as `.xml` without the `.gz` ending. That would
however only result in larger file sizes when saved and would give no
extra benefit. It would also lead to non-compatible changes.)