Performance optimize onTriggerSearch
Previous to this fix, onTriggerSearch triggered 28 times during loading
and edit page with two search blocks in it. After the fix, it's just 2
times.

The fix is that the deep comparison condition for the query also has to
be applied in withSearch.jsx. Without that, the deep comparison in
SearchBlockEdit.jsx has no effect. This is because the onTriggerSearch
function will be regenerated each time there is an identity change in
data.query anyway. So the hook guard in SearchBlockEdit will be
triggered anyway because onTriggerSearch has changed.

In addition, I personally find it better to use JSON.stringify for the
deep comparison and skip using the useDeepCompareEffect hook. This uses
dequal under the hood which is capable to compare a lot of other data
types, but with pure json data it has no advantage to a string
comparison with JSON.stringify. (That said it would also work with
useDeepCompareEffect but perhaps a bit slower.)