Remove SearchableText from the query only if there is a searchtext
Previously, the SearchableText was removed within normalizeState, even
if there was no searchText. This is wrong, the searchText should only
replace SearchableText condition if searchText is provided.

In the edit page this caused the following, in case a SearchableText
is specified as a condition to the block:

- Upon loading the edit page
- First the search blocks were loaded with correct results
- Then a second set of query was issued, but with the SearchableText
  condition stripped out. This means that the query result was wrong.

The current fix fixes the double querying, and the faulty condition in
the second query that results in faulty results.

The current fix still does not address the fact that eventually the
searchText should be a distinct facet from SearchableText, and the two
conditions could be combined, in comparison to the current state, when
one overrides the other.