Fix condition for view translation, for missing view case #5351
Fixes clicking on three dots, which causes an error in case of layout is missing,

It's unclear what causes the layout to be missing. However
from the asyncconnect code it's obvious that this is a
case to be prepared for, therefore we must be tolerant
in the case the layout to be unset, instead of letting
the translations engine fail - which caused this error.