Fix cyprees test "As editor, I can unlock a locked page"
Sometimes this test fails because it does not find the alert with the
message that the content is locked. What may be happening is that the
test runs too quickly, and Plone does not consider that the content has
been locked. Then we wait a bit for Plone to consider it locked.