Fixes imports to avoid type errors in addons tests
When we shadow a module in an addon and this module is imported
relatively in some .ts file used in the addon tests, we receive the
error:

Could not find a declaration file for module './Module'. '../Module.jsx'
implicitly has an 'any' type

So we use absolute import in .ts files to avoid this problem.