Remove "Generate tabs for items other than folders" field from navigation control panel
The help for this field says:

"By default, any content item in the root of the portal will appear as a
tab. If you turn this option off, only folders will be shown. This only
has an effect if 'automatically generate tabs' is enabled."

This doesn't make sense in Volto. Volto doesn't display item as tab