Stop generating sitemap.xml.gz (#6561)
We generate sitemap-index.xml which is also put correctly into robots.txt. However we still provide the old sitemap.xml.gz.

But, while this serves no purpose in addition to the index, it can cause problems if Google somehow parses it (for example it's submitted to the Google Search console). For this reason, we  stop providing sitemap.xml.gz.

- also serve the batched sitemap under the old name sitemap.xml.gz
- update comment in robots.txt