Fix Cypress "Test Yearly recurrence" test
When we create an Event, the times of start date and times of end date
are automatically filled in based on the current time. However, the test
only changed the start and end date, leaving the time that was
automatically filled in. So, depending on the time the test was run, the
number of recursive events calculated could be 12 or 13. Depending on
the time, the last day was not considered, as it did not include the
entire event, up to the end time.

The error occurred:

Timed out retrying after 4000ms: Not enough elements found. Found '12',
expected '13'.

Now we fill in the times so that there are always 13 days of events.